Add initial Bazel Module support for tink-cc and Kokoro test script
Bazel Modules [1] is the new external dependency management system in Bazel.
As of Bazel 6.0.0 bzlmod must be enabled with `--enable_bzlmod`[2].

For certain dependencies, namely absl, wycheproof, rapidjson and boringssl, we rely on module extensions, since the versions available in the Bazel Central Registry [2] are older than the ones Tink currently uses.

[1] https://bazel.build/versions/6.0.0/build/bzlmod
[2] https://github.com/bazelbuild/bazel/releases/tag/6.0.0
[3] https://bazel.build/versions/6.0.0/build/bzlmod#bazel-central-registry

PiperOrigin-RevId: 509457563
Change-Id: I32f3d40ce81a126f7d3a5bb03470c5a56342b00c