Document how Tink handles associatedData='' in AES SIV.
AES SIV allows the user to import a list of associated datas, but Tink only uses a single associated data in its interface. When the user inputs an empty associated data byte sequence, Tink will handle this as a list of one empty byte sequence, and not as an empty list.

PiperOrigin-RevId: 527491521
Change-Id: Iedfcbee204db27856d5891027a57709ef8a5c754