Restrict KMS envelope AEAD to only use Tink AEAD key types as DEK.
KMS Envelope AEAD is a simple way to encrypt data. So using custom
implementations of AEAD is not needed for this.

This also prevents the user from using the envelope encryption key type as DEK by accident.

This change will help us simplify our implementation of KMS Envelope AEAD,
because we will be able to remove the dependency on Tink's Registry.

PiperOrigin-RevId: 543969066
Change-Id: I6971d3e7f38e7f746a2278606e7375fc6ef1a216