Add JSON parsing tests with tailing commas.
This is not allowed by the spec, see https://www.json.org/json-en.html.

The rapidjson parser currently used to parse keysets does not allow this, but the proto JSON parser used for JWTs does allow it.

The proto parser doesn't have the option to not allow this, so there is not easy way to fix this.

PiperOrigin-RevId: 642986258
Change-Id: I962209a9ebb3d6220b2216d35d8a8185efe382c6