Add a class "ParsingState".
Previously, when parsing we would maintain a "string_view&" which contains the remaining data to parse. We change this to maintain a "ParsingState&".

This allows us later to include CRC values in the parsingstate. Besides this, i actually think this makes the code easier to read (since a ParsingState more clearly communicates the purpose than a string_view).

PiperOrigin-RevId: 675970220
Change-Id: I40b2409ad3f9d86b4cde989c16d028d29c6f9572