When parsing a submessage do not create a new state. Instead, obtain a "submessage state"
The problem with the current approach is that we don't do proper CRC calculation in inner messages. We simply created a new state, when then is without CRC maintenance. Instead, we need to have a state with CRC maintenance.

Unfortunately, I don't have a good way to test this (except that once we finally enable CRC parsing fully we wouldn't be able to use SecretDataWithCRC fields in submessages since they require maintenance).

PiperOrigin-RevId: 678576584
Change-Id: I25d1155355e718e58ec099483ebf2513b8797309