Rename certain key symbols locally to have "proto" in the name.
For example, we have two symbols EcdsaPrivateKey. One lives in google::crypto::tink::EcdsaPublicKey -- the proto. The other lives in crypto::tink::EcdsaPublicKey.

This isn't a problem a priori, but if we write "using google::crypto::tink::EcdsaPublicKey" while in namespace crypto::tink it becomes a problem because the declaration states that we create a new symbol "crypto::tink::EcdsaPublicKey" which is an alias for "google::crypto::tink::EcdsaPublicKey", and now we have two symbols.

In many cases we are additionally in an anonymous namespace.  Then the above problem doesn't happen but it will still be ambiguous. So it's best to avoid this problem everywhere.

PiperOrigin-RevId: 680989066
Change-Id: I8858297d4bcc47931f34ce07c7ab2e1f94614f3e