Refactor the getPublicKeysetHandle function to avoid using key managers if an entry for the private key is available or possible to create.
This also avoids an unnecessary loop through the keyset to create the entires. However, creating an entry from a key reference is not possible at the moment, so we would probably need to implement Clone() to avoid serializing and parsing again.

Falling back to using the registry/ key managers is still needed for some use cases.

PiperOrigin-RevId: 681435780
Change-Id: I9fb6c4bf7e64dd613241bb796d24ea525bab9142