Inline the DummyStatefulMac into the only place it is used.
This removes DummyStatefulMac from test_util. This was never intended to be public API. If this affects you, please file an issue in github.com/tink-crypto/tink-cc so we better understand the usage. We recommend to re-implement this in your code instead.

I want to remove StatefulMac from the public API because it would be beneficial if Finalize would return a SecretData: if one implements VerifyMac one should not leak the correct tag.

PiperOrigin-RevId: 691002448
Change-Id: I3b5bee9e65d35c2af2cd378cbae56bd8ea2324f8