Add a method "FinalizeAsSecretData" to StatefulMac.
Because StatefulMac is used to compute ChunkedMac, we need to core dump protect the tag. This means we cannot keep it in a std::string, and I will migrate all usages. I think it's too much work for a single CL, so I will first migrate all calls, then all implementations separately.

PiperOrigin-RevId: 693338026
Change-Id: I2b35288865a48216daf9567f70734bed116e9815