Change the semantics of GetSerializedSizeIncludingTag.
In the proto format, one can sometimes omit fields. For example, if a int has value 0 and the default is defined to be 0, one doesn't have to serialize it.

Previously, GetSerializedSizeIncludingTag would return the size of the serialization including tag *assuming it needs to be serialized*. After this CL, it returns 0 in case it does not need to be serialized.

This will allow us to also remove RequiresSerialization from the field API in the future, making things simpler.

This change is safe since we only use GetSerializedSizeIncludingTag in this one place at the moment. It was introduced in the previous CL.

PiperOrigin-RevId: 705049086
Change-Id: Idcec9ffa6afa8f7e8b07ff1de2498f3bcb6672c6