Change the test of MessageField by testing SerializeWithTagInto and GetSerializedSizeIncludingTag, instead of SerializeInto and GetSerializedSize.
Make GetSerializedSize and SerializeInto protected to ensure we have all instances. In upcoming CLs I will inline those into SerializeWithTagInto and GetSerializedSizeIncludingTag.

After this test was written, we have introduced the testing library "FieldWithNumber". "FieldWithNumber(1).IsVarint(10)" will return the string which a field with number 1 which is a varint with value 10 has.

PiperOrigin-RevId: 705096714
Change-Id: I1be60d3c89fabf015f025ab1263dcf2c6fc77e0d