Remove SerializeInto and GetSerializedSize.
The field interface (in proto_parser_fields) loses the methods SerializeInto and GetSerializedSize. Instead, subclasses need to directly implement SerializeWithTagInto and GetSerializedSizeIncludingTag.

To implement this, we can follow a simple recipe.

For SerializeInto, we prepend the following to the remaining code:
```
    if (!RequiresSerialization(values)) {
      return absl::OkStatus();
    }
    absl::Status status = SerializeWireTypeAndFieldNumber(
        GetWireType(), GetFieldNumber(), serialization_state);
    if (!status.ok()) {
      return status;
    }
```

For GetSerializedSize we prepend:

```
    if (!RequiresSerialization(values)) {
      return 0;
    }
```

and also add `WireTypeAndFieldNumberLength(GetWireType(), GetFieldNumber())` to the result.

PiperOrigin-RevId: 705786676
Change-Id: I4f2f83d659165e93b318d36263b7f127a54777e7